+++ libs/tendermint/consensus/consensus_commit.go
@@ -1,0 +1,356 @@
+package consensus
+
+import (
+	"encoding/hex"
+	"fmt"
+	"github.com/okex/exchain/libs/system/trace"
+	cstypes "github.com/okex/exchain/libs/tendermint/consensus/types"
+	"github.com/okex/exchain/libs/tendermint/libs/fail"
+	tmos "github.com/okex/exchain/libs/tendermint/libs/os"
+	sm "github.com/okex/exchain/libs/tendermint/state"
+	"github.com/okex/exchain/libs/tendermint/types"
+)
+
+
+func (cs *State) traceDump() {
+	if cs.Logger == nil {
+		return
+	}
+
+	trace.GetElapsedInfo().AddInfo(trace.CommitRound, fmt.Sprintf("%d", cs.CommitRound))
+	trace.GetElapsedInfo().AddInfo(trace.Round, fmt.Sprintf("%d", cs.Round))
+	trace.GetElapsedInfo().AddInfo(trace.BlockParts, fmt.Sprintf("%d|%d|%d/%d",
+		cs.bt.droppedDue2WrongHeight,
+		cs.bt.droppedDue2NotExpected,
+		cs.bt.droppedDue2NotAdded,
+		cs.bt.totalParts,
+	))
+
+	trace.GetElapsedInfo().AddInfo(trace.Produce, cs.trc.Format())
+	trace.GetElapsedInfo().Dump(cs.Logger.With("module", "main"))
+	cs.trc.Reset()
+}
+
+// Enter: +2/3 precommits for block
+func (cs *State) enterCommit(height int64, commitRound int) {
+	logger := cs.Logger.With("height", height, "commitRound", commitRound)
+
+	if cs.Height != height || cstypes.RoundStepCommit <= cs.Step {
+		logger.Debug(fmt.Sprintf(
+			"enterCommit(%v/%v): Invalid args. Current step: %v/%v/%v",
+			height,
+			commitRound,
+			cs.Height,
+			cs.Round,
+			cs.Step))
+		return
+	}
+
+	cs.initNewHeight()
+	cs.trc.Pin("%s-%d-%d", "Commit", cs.Round, commitRound)
+
+	logger.Info(fmt.Sprintf("enterCommit(%v/%v). Current: %v/%v/%v", height, commitRound, cs.Height, cs.Round, cs.Step))
+
+	defer func() {
+		// Done enterCommit:
+		// keep cs.Round the same, commitRound points to the right Precommits set.
+		cs.updateRoundStep(cs.Round, cstypes.RoundStepCommit)
+		cs.CommitRound = commitRound
+		cs.newStep()
+
+		// Maybe finalize immediately.
+		cs.tryFinalizeCommit(height)
+	}()
+
+	blockID, ok := cs.Votes.Precommits(commitRound).TwoThirdsMajority()
+	if !ok {
+		panic("RunActionCommit() expects +2/3 precommits")
+	}
+
+	// The Locked* fields no longer matter.
+	// Move them over to ProposalBlock if they match the commit hash,
+	// otherwise they'll be cleared in updateToState.
+	if cs.LockedBlock.HashesTo(blockID.Hash) {
+		logger.Info("Commit is for locked block. Set ProposalBlock=LockedBlock", "blockHash", blockID.Hash)
+		cs.ProposalBlock = cs.LockedBlock
+		cs.ProposalBlockParts = cs.LockedBlockParts
+	}
+
+	// If we don't have the block being committed, set up to get it.
+	if !cs.ProposalBlock.HashesTo(blockID.Hash) {
+		if !cs.ProposalBlockParts.HasHeader(blockID.PartsHeader) {
+			logger.Info(
+				"Commit is for a block we don't know about. Set ProposalBlock=nil",
+				"proposal",
+				cs.ProposalBlock.Hash(),
+				"commit",
+				blockID.Hash)
+			// We're getting the wrong block.
+			// Set up ProposalBlockParts and keep waiting.
+			cs.ProposalBlock = nil
+			cs.Logger.Info("enterCommit proposalBlockPart reset ,because of mismatch hash,",
+				"origin", hex.EncodeToString(cs.ProposalBlockParts.Hash()), "after", blockID.Hash)
+			cs.ProposalBlockParts = types.NewPartSetFromHeader(blockID.PartsHeader)
+			cs.eventBus.PublishEventValidBlock(cs.RoundStateEvent())
+			cs.evsw.FireEvent(types.EventValidBlock, &cs.RoundState)
+		}
+		// else {
+		// We just need to keep waiting.
+		// }
+	}
+}
+
+// If we have the block AND +2/3 commits for it, finalize.
+func (cs *State) tryFinalizeCommit(height int64) {
+	logger := cs.Logger.With("height", height)
+
+	if cs.Height != height {
+		panic(fmt.Sprintf("tryFinalizeCommit() cs.Height: %v vs height: %v", cs.Height, height))
+	}
+
+	blockID, ok := cs.Votes.Precommits(cs.CommitRound).TwoThirdsMajority()
+	if !ok || len(blockID.Hash) == 0 {
+		logger.Error("Attempt to finalize failed. There was no +2/3 majority, or +2/3 was for <nil>.")
+		return
+	}
+	if !cs.ProposalBlock.HashesTo(blockID.Hash) {
+		// TODO: this happens every time if we're not a validator (ugly logs)
+		// TODO: ^^ wait, why does it matter that we're a validator?
+		logger.Info(
+			"Attempt to finalize failed. We don't have the commit block.",
+			"proposal-block",
+			cs.ProposalBlock.Hash(),
+			"commit-block",
+			blockID.Hash)
+		return
+	}
+
+	//	go
+	cs.finalizeCommit(height)
+}
+
+// Increment height and goto cstypes.RoundStepNewHeight
+func (cs *State) finalizeCommit(height int64) {
+	if cs.Height != height || cs.Step != cstypes.RoundStepCommit {
+		cs.Logger.Debug(fmt.Sprintf(
+			"finalizeCommit(%v): Invalid args. Current step: %v/%v/%v",
+			height,
+			cs.Height,
+			cs.Round,
+			cs.Step))
+		return
+	}
+
+	blockID, ok := cs.Votes.Precommits(cs.CommitRound).TwoThirdsMajority()
+	block, blockParts := cs.ProposalBlock, cs.ProposalBlockParts
+
+	if !ok {
+		panic(fmt.Sprintf("Cannot finalizeCommit, commit does not have two thirds majority"))
+	}
+	if !blockParts.HasHeader(blockID.PartsHeader) {
+		panic(fmt.Sprintf("Expected ProposalBlockParts header to be commit header"))
+	}
+	if !block.HashesTo(blockID.Hash) {
+		panic(fmt.Sprintf("Cannot finalizeCommit, ProposalBlock does not hash to commit hash"))
+	}
+	if err := cs.blockExec.ValidateBlock(cs.state, block); err != nil {
+		panic(fmt.Sprintf("+2/3 committed an invalid block: %v", err))
+	}
+
+	cs.Logger.Info("Finalizing commit of block with N txs",
+		"height", block.Height,
+		"hash", block.Hash(),
+		"root", block.AppHash,
+		"N", len(block.Txs))
+	cs.Logger.Info(fmt.Sprintf("%v", block))
+
+	fail.Fail() // XXX
+
+	// Save to blockStore.
+	if cs.blockStore.Height() < block.Height {
+		// NOTE: the seenCommit is local justification to commit this block,
+		// but may differ from the LastCommit included in the next block
+		precommits := cs.Votes.Precommits(cs.CommitRound)
+		seenCommit := precommits.MakeCommit()
+		cs.blockStore.SaveBlock(block, blockParts, seenCommit)
+	} else {
+		// Happens during replay if we already saved the block but didn't commit
+		cs.Logger.Info("Calling finalizeCommit on already stored block", "height", block.Height)
+	}
+
+	fail.Fail() // XXX
+
+	// Write EndHeightMessage{} for this height, implying that the blockstore
+	// has saved the block.
+	//
+	// If we crash before writing this EndHeightMessage{}, we will recover by
+	// running ApplyBlock during the ABCI handshake when we restart.  If we
+	// didn't save the block to the blockstore before writing
+	// EndHeightMessage{}, we'd have to change WAL replay -- currently it
+	// complains about replaying for heights where an #ENDHEIGHT entry already
+	// exists.
+	//
+	// Either way, the State should not be resumed until we
+	// successfully call ApplyBlock (ie. later here, or in Handshake after
+	// restart).
+	endMsg := EndHeightMessage{height}
+	if err := cs.wal.WriteSync(endMsg); err != nil { // NOTE: fsync
+		panic(fmt.Sprintf("Failed to write %v msg to consensus wal due to %v. Check your FS and restart the node",
+			endMsg, err))
+	}
+
+	fail.Fail() // XXX
+
+	// Create a copy of the state for staging and an event cache for txs.
+	stateCopy := cs.state.Copy()
+
+	// Execute and commit the block, update and save the state, and update the mempool.
+	// NOTE The block.AppHash wont reflect these txs until the next block.
+
+	var err error
+	var retainHeight int64
+
+	cs.trc.Pin("%s-%d", trace.RunTx, cs.Round)
+	stateCopy, retainHeight, err = cs.blockExec.ApplyBlock(
+		stateCopy,
+		types.BlockID{Hash: block.Hash(), PartsHeader: blockParts.Header()},
+		block)
+	if err != nil {
+		cs.Logger.Error("Error on ApplyBlock. Did the application crash? Please restart tendermint", "err", err)
+		err := tmos.Kill()
+		if err != nil {
+			cs.Logger.Error("Failed to kill this process - please do so manually", "err", err)
+		}
+		return
+	}
+
+	fail.Fail() // XXX
+
+	// Prune old heights, if requested by ABCI app.
+	if retainHeight > 0 {
+		pruned, err := cs.pruneBlocks(retainHeight)
+		if err != nil {
+			cs.Logger.Error("Failed to prune blocks", "retainHeight", retainHeight, "err", err)
+		} else {
+			cs.Logger.Info("Pruned blocks", "pruned", pruned, "retainHeight", retainHeight)
+		}
+	}
+
+	// must be called before we update state
+	cs.recordMetrics(height, block)
+
+	// NewHeightStep!
+	cs.updateToState(stateCopy)
+
+	fail.Fail() // XXX
+
+	// Private validator might have changed it's key pair => refetch pubkey.
+	if err := cs.updatePrivValidatorPubKey(); err != nil {
+		cs.Logger.Error("Can't get private validator pubkey", "err", err)
+	}
+
+	cs.trc.Pin("Waiting")
+	// cs.StartTime is already set.
+	// Schedule Round0 to start soon.
+	cs.scheduleRound0(&cs.RoundState)
+
+	// By here,
+	// * cs.Height has been increment to height+1
+	// * cs.Step is now cstypes.RoundStepNewHeight
+	// * cs.StartTime is set to when we will start round0.
+}
+
+// Updates State and increments height to match that of state.
+// The round becomes 0 and cs.Step becomes cstypes.RoundStepNewHeight.
+func (cs *State) updateToState(state sm.State) {
+	// Do not consider this situation that the consensus machine was stopped
+	// when the fast-sync mode opens. So remove it!
+	//if cs.CommitRound > -1 && 0 < cs.Height && cs.Height != state.LastBlockHeight {
+	//	panic(fmt.Sprintf("updateToState() expected state height of %v but found %v",
+	//		cs.Height, state.LastBlockHeight))
+	//}
+	//if !cs.state.IsEmpty() && cs.state.LastBlockHeight+1 != cs.Height {
+	//	// This might happen when someone else is mutating cs.state.
+	//	// Someone forgot to pass in state.Copy() somewhere?!
+	//	panic(fmt.Sprintf("Inconsistent cs.state.LastBlockHeight+1 %v vs cs.Height %v",
+	//		cs.state.LastBlockHeight+1, cs.Height))
+	//}
+
+	// If state isn't further out than cs.state, just ignore.
+	// This happens when SwitchToConsensus() is called in the reactor.
+	// We don't want to reset e.g. the Votes, but we still want to
+	// signal the new round step, because other services (eg. txNotifier)
+	// depend on having an up-to-date peer state!
+	if !cs.state.IsEmpty() && (state.LastBlockHeight <= cs.state.LastBlockHeight) {
+		cs.Logger.Info(
+			"Ignoring updateToState()",
+			"newHeight",
+			state.LastBlockHeight+1,
+			"oldHeight",
+			cs.state.LastBlockHeight+1)
+		cs.newStep()
+		return
+	}
+
+	// Reset fields based on state.
+	validators := state.Validators
+	lastPrecommits := (*types.VoteSet)(nil)
+	if cs.CommitRound > -1 && cs.Votes != nil {
+		if !cs.Votes.Precommits(cs.CommitRound).HasTwoThirdsMajority() {
+			panic("updateToState(state) called but last Precommit round didn't have +2/3")
+		}
+		lastPrecommits = cs.Votes.Precommits(cs.CommitRound)
+	}
+
+	// Next desired block height
+	height := state.LastBlockHeight + 1
+
+	// RoundState fields
+	cs.updateHeight(height)
+	cs.updateRoundStep(0, cstypes.RoundStepNewHeight)
+	cs.traceDump()
+	cs.bt.reset(height)
+
+	cs.Validators = validators
+	cs.Proposal = nil
+	cs.ProposalBlock = nil
+	cs.ProposalBlockParts = nil
+	cs.LockedRound = -1
+	cs.LockedBlock = nil
+	cs.LockedBlockParts = nil
+	cs.ValidRound = -1
+	cs.ValidBlock = nil
+	cs.ValidBlockParts = nil
+	cs.Votes = cstypes.NewHeightVoteSet(state.ChainID, height, validators)
+	cs.CommitRound = -1
+	cs.LastCommit = lastPrecommits
+	cs.LastValidators = state.LastValidators
+	cs.TriggeredTimeoutPrecommit = false
+	cs.state = state
+
+	// Finally, broadcast RoundState
+	cs.newStep()
+}
+
+
+func (cs *State) updateHeight(height int64) {
+	cs.metrics.Height.Set(float64(height))
+	cs.Height = height
+}
+
+
+func (cs *State) pruneBlocks(retainHeight int64) (uint64, error) {
+	base := cs.blockStore.Base()
+	if retainHeight <= base {
+		return 0, nil
+	}
+	pruned, err := cs.blockStore.PruneBlocks(retainHeight)
+	if err != nil {
+		return 0, fmt.Errorf("failed to prune block store: %w", err)
+	}
+	err = sm.PruneStates(cs.blockExec.DB(), base, retainHeight)
+	if err != nil {
+		return 0, fmt.Errorf("failed to prune state database: %w", err)
+	}
+	return pruned, nil
+}
