+++ libs/tendermint/consensus/consensus_precommit.go
@@ -1,0 +1,155 @@
+package consensus
+
+import (
+	"fmt"
+	cstypes "github.com/okex/exchain/libs/tendermint/consensus/types"
+	"github.com/okex/exchain/libs/tendermint/libs/automation"
+	"github.com/okex/exchain/libs/tendermint/types"
+)
+
+// Enter: `timeoutPrevote` after any +2/3 prevotes.
+// Enter: `timeoutPrecommit` after any +2/3 precommits.
+// Enter: +2/3 precomits for block or nil.
+// Lock & precommit the ProposalBlock if we have enough prevotes for it (a POL in this round)
+// else, unlock an existing lock and precommit nil if +2/3 of prevotes were nil,
+// else, precommit nil otherwise.
+func (cs *State) enterPrecommit(height int64, round int) {
+	logger := cs.Logger.With("height", height, "round", round)
+
+	if cs.Height != height || round < cs.Round || (cs.Round == round && cstypes.RoundStepPrecommit <= cs.Step) {
+		logger.Debug(fmt.Sprintf(
+			"enterPrecommit(%v/%v): Invalid args. Current step: %v/%v/%v",
+			height,
+			round,
+			cs.Height,
+			cs.Round,
+			cs.Step))
+		return
+	}
+
+	cs.initNewHeight()
+	cs.trc.Pin("Precommit-%d", round)
+
+	logger.Info(fmt.Sprintf("enterPrecommit(%v/%v). Current: %v/%v/%v", height, round, cs.Height, cs.Round, cs.Step))
+
+	defer func() {
+		// Done enterPrecommit:
+		cs.updateRoundStep(round, cstypes.RoundStepPrecommit)
+		cs.newStep()
+	}()
+
+	if automation.PrecommitNil(height, round) {
+		cs.signAddVote(types.PrecommitType, nil, types.PartSetHeader{})
+		return
+	}
+
+	// check for a polka
+	blockID, ok := cs.Votes.Prevotes(round).TwoThirdsMajority()
+
+	// If we don't have a polka, we must precommit nil.
+	if !ok {
+		if cs.LockedBlock != nil {
+			logger.Info("enterPrecommit: No +2/3 prevotes during enterPrecommit while we're locked. Precommitting nil")
+		} else {
+			logger.Info("enterPrecommit: No +2/3 prevotes during enterPrecommit. Precommitting nil.")
+		}
+		cs.signAddVote(types.PrecommitType, nil, types.PartSetHeader{})
+		return
+	}
+
+	// At this point +2/3 prevoted for a particular block or nil.
+	cs.eventBus.PublishEventPolka(cs.RoundStateEvent())
+
+	// the latest POLRound should be this round.
+	polRound, _ := cs.Votes.POLInfo()
+	if polRound < round {
+		panic(fmt.Sprintf("This POLRound should be %v but got %v", round, polRound))
+	}
+
+	// +2/3 prevoted nil. Unlock and precommit nil.
+	if len(blockID.Hash) == 0 {
+		if cs.LockedBlock == nil {
+			logger.Info("enterPrecommit: +2/3 prevoted for nil.")
+		} else {
+			logger.Info("enterPrecommit: +2/3 prevoted for nil. Unlocking")
+			cs.LockedRound = -1
+			cs.LockedBlock = nil
+			cs.LockedBlockParts = nil
+			cs.eventBus.PublishEventUnlock(cs.RoundStateEvent())
+		}
+		cs.signAddVote(types.PrecommitType, nil, types.PartSetHeader{})
+		return
+	}
+
+	// At this point, +2/3 prevoted for a particular block.
+
+	// If we're already locked on that block, precommit it, and update the LockedRound
+	if cs.LockedBlock.HashesTo(blockID.Hash) {
+		logger.Info("enterPrecommit: +2/3 prevoted locked block. Relocking")
+		cs.LockedRound = round
+		cs.eventBus.PublishEventRelock(cs.RoundStateEvent())
+		cs.signAddVote(types.PrecommitType, blockID.Hash, blockID.PartsHeader)
+		return
+	}
+
+	// If +2/3 prevoted for proposal block, stage and precommit it
+	if cs.ProposalBlock.HashesTo(blockID.Hash) {
+		logger.Info("enterPrecommit: +2/3 prevoted proposal block. Locking", "hash", blockID.Hash)
+		// Validate the block.
+		if err := cs.blockExec.ValidateBlock(cs.state, cs.ProposalBlock); err != nil {
+			panic(fmt.Sprintf("enterPrecommit: +2/3 prevoted for an invalid block: %v", err))
+		}
+		cs.LockedRound = round
+		cs.LockedBlock = cs.ProposalBlock
+		cs.LockedBlockParts = cs.ProposalBlockParts
+		cs.eventBus.PublishEventLock(cs.RoundStateEvent())
+		cs.signAddVote(types.PrecommitType, blockID.Hash, blockID.PartsHeader)
+		return
+	}
+
+	// There was a polka in this round for a block we don't have.
+	// Fetch that block, unlock, and precommit nil.
+	// The +2/3 prevotes for this round is the POL for our unlock.
+	// TODO: In the future save the POL prevotes for justification.
+	cs.LockedRound = -1
+	cs.LockedBlock = nil
+	cs.LockedBlockParts = nil
+	if !cs.ProposalBlockParts.HasHeader(blockID.PartsHeader) {
+		cs.ProposalBlock = nil
+		cs.ProposalBlockParts = types.NewPartSetFromHeader(blockID.PartsHeader)
+	}
+	cs.eventBus.PublishEventUnlock(cs.RoundStateEvent())
+	cs.signAddVote(types.PrecommitType, nil, types.PartSetHeader{})
+}
+
+// Enter: any +2/3 precommits for next round.
+func (cs *State) enterPrecommitWait(height int64, round int) {
+	logger := cs.Logger.With("height", height, "round", round)
+
+	if cs.Height != height || round < cs.Round || (cs.Round == round && cs.TriggeredTimeoutPrecommit) {
+		logger.Debug(
+			fmt.Sprintf(
+				"enterPrecommitWait(%v/%v): Invalid args. "+
+					"Current state is Height/Round: %v/%v/, TriggeredTimeoutPrecommit:%v",
+				height, round, cs.Height, cs.Round, cs.TriggeredTimeoutPrecommit))
+		return
+	}
+
+	cs.initNewHeight()
+	cs.trc.Pin("PrecommitWait-%d", round)
+
+	if !cs.Votes.Precommits(round).HasTwoThirdsAny() {
+		panic(fmt.Sprintf("enterPrecommitWait(%v/%v), but Precommits does not have any +2/3 votes", height, round))
+	}
+	logger.Info(fmt.Sprintf("enterPrecommitWait(%v/%v). Current: %v/%v/%v", height, round, cs.Height, cs.Round, cs.Step))
+
+	defer func() {
+		// Done enterPrecommitWait:
+		cs.TriggeredTimeoutPrecommit = true
+		cs.newStep()
+	}()
+
+	// Wait for some more precommits; enterNewRound
+	cs.scheduleTimeout(cs.config.Precommit(round), height, round, cstypes.RoundStepPrecommitWait)
+
+}
